+++ Assets/Scripts/Sound/EnemyEngineSound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a50821c734546f4bb4e53375ba07f41
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
