+++ UserSettings/EditorUserSettings.asset
@@ -12,28 +12,28 @@ EditorUserSettings:
       value: 0206065207035c59095d5d2747225c4442161b2f7e797136787e1b37e1b6616e
       flags: 0
     RecentlyUsedSceneGuid-2:
-      value: 5055555201570f0f595f5b2611215c4413164b2e2a7f7e3329714862b4e53668
-      flags: 0
-    RecentlyUsedSceneGuid-3:
       value: 55020d0500505f035b0a5d75127b5d4443161b7f782925342b791c30b5b76c3c
       flags: 0
-    RecentlyUsedSceneGuid-4:
-      value: 52030d0057510f025e5d54264225064415161a7e7c2c7e62782b4f37e0b7326b
-      flags: 0
-    RecentlyUsedSceneGuid-5:
+    RecentlyUsedSceneGuid-3:
       value: 5004065f5c545f5a5457087414255a4414154f7a757d74672f284466e3e13668
       flags: 0
-    RecentlyUsedSceneGuid-6:
+    RecentlyUsedSceneGuid-4:
       value: 550500055356515d0b580f7216200e44154e1c7d2d7a2264782c4465e0b16d6d
       flags: 0
-    RecentlyUsedSceneGuid-7:
+    RecentlyUsedSceneGuid-5:
       value: 01020d040003515a095d0d2449760f4446164129292d236374701935e6e1636c
       flags: 0
-    RecentlyUsedSceneGuid-8:
+    RecentlyUsedSceneGuid-6:
       value: 5253565e5c570d025b5d0d7749200b44154e1e79787d2367782f4432b6b0323a
       flags: 0
-    RecentlyUsedSceneGuid-9:
+    RecentlyUsedSceneGuid-7:
+      value: 52030d0057510f025e5d54264225064415161a7e7c2c7e62782b4f37e0b7326b
+      flags: 0
+    RecentlyUsedSceneGuid-8:
       value: 5504030503055c585b5b5b76407a5c44104e487e7c7d71322e7f486bb1e36461
+      flags: 0
+    RecentlyUsedSceneGuid-9:
+      value: 5055555201570f0f595f5b2611215c4413164b2e2a7f7e3329714862b4e53668
       flags: 0
     UnityRemoteDevice:
       value: 2d5e5a03
